requirements.txt
alabaster==0.7.12
async==0.6.2
Babel==2.7.0
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
cycler==0.10.0
dj-database-url==0.5.0
Django==2.2.1
django-filter==2.1.0
django-heroku==0.3.1
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
docutils==0.14
gunicorn==19.9.0
idna==2.8
imagesize==1.1.0
Jinja2==2.10.1
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mod-wsgi==4.6.5
mysql-connector-python==8.0.16
mysqlclient==1.4.2.post1
nltk==3.4.1
numpy==1.16.4
oauthlib==3.0.1
packaging==19.0
protobuf==3.8.0
psycopg2==2.8.3
Pygments==2.4.2
PyJWT==1.7.1
pyparsing==2.4.0
PySocks==1.7.0
python-dateutil==2.8.0
python-firebase==1.2
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
snowballstemmer==1.2.1
Sphinx==2.1.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
textblob==0.15.3
tweepy==3.7.0
urllib3==1.25.3
whitenoise==4.1.2
